Former-commit-id: 430e75f5dae0dff58c3a56d827409f4161939207
Former-commit-id: e33f918e9839e77394a662110442289df3b7ee34
Former-commit-id: 2b94c32bea4e9a4335edd3587f5ca7a8156657cc

+++ src/ImageProcessor/Samplers/Resamplers/Lanczos3Resampler.cs
@@ -12,10 +12,10 @@ namespace ImageProcessor.Samplers
     public class Lanczos3Resampler : IResampler
     {
         /// <inheritdoc/>
-        public double Radius => 3;
+        public float Radius => 3;
 
         /// <inheritdoc/>
-        public double GetValue(double x)
+        public float GetValue(float x)
         {
             if (x < 0)
             {
